Update geturl() function error handling
The lodgeit engine was the last engine in wgetpaste that needed the
third branch in geturl().  With its removal, that branch can now be
updated to issue an error if REGEX_LOC_$ENGINE or REGEX_URL_$ENGINE is
not defined.

Requiring those variables to be defined makes geturl() less likely to
grab the incorrect URL.

Fixes bug #10